Module Camllib

module DHashhe : sig ... end

Two-way hashtable between two data types

module DMappe : sig ... end

Two-way map between two ordered data types

module FGraph : sig ... end

Directed graphs, functional API, with two-way information maintained,

module FGraph1 : sig ... end

Directed graphs, functional API, with one-way information maintained,

module Hashhe : sig ... end

Hash tables and hash functions (extension of standard library module)

module HashheI : sig ... end

Hashhe specialized for keys of type int

module HashheIB : sig ... end

Hashhe specialized for keys of type (int,bool)

module HashheS : sig ... end

Hashhe specialized for keys of type string

module Ilist : sig ... end

Imbricated lists

module Mappe : sig ... end

Association tables over ordered types (extension of standard library module and polymorphic variant)

module MappeI : sig ... end

Mappe specialized for keys of type int

module MappeS : sig ... end

Mappe specialized for keys of type string

module MultiSetList : sig ... end

Multisets implemented with lists

module PDHashhe : sig ... end

Two-way hashtable between two data types, parametrized polymorpphic version

module PDMappe : sig ... end

Two-way map between two ordered data types, parametrized polymorphic version

module PHashhe : sig ... end

Hash tables, parametrized polymorphic version

module PMappe : sig ... end

Association tables over ordered types, parametrized polymorphic version

module PSHGraph : sig ... end

Oriented hypergraphs, parametrized polymorphic version.

module PSette : sig ... end

Sets over ordered types, parametrized polymorphic version

module Parse : sig ... end

Linking lexer and parser function, with (basic) error messages

module Print : sig ... end

Printing functions using module Format

module Rational : sig ... end

Rational numbers

module SHGraph : sig ... end

Oriented hypergraphs

module SetList : sig ... end

Sets over totally ordered type with lists. All operations over sets are purely applicative (no side-effects).

module Sette : sig ... end

Sets over ordered types (extension of standard library module and polymorphic variant)

module SetteI : sig ... end

Sette specialized for keys of type int

module SetteS : sig ... end

Sette specialized for keys of type string

module Statistic : sig ... end

Standard statistics functions

module Symbol : sig ... end

Symbol table, for string management

module Time : sig ... end

Small module to compute the duration of computations

module Union : sig ... end

Union-find Abstract Data Types