Module Camllib.DHashhe

Two-way hashtable between two data types

type ('a, 'b) t = {

xy : ('a, 'b) Hashhe.t;

yx : ('b, 'a) Hashhe.t;

}

The type of two-way hashtables, meant to be abstract

val hashx : ('a, 'b) t -> ('a, 'b) Hashhe.t

val hashy : ('a, 'b) t -> ('b, 'a) Hashhe.t

Return the correspondance hashtable resp. from x to y and from y to x. Never modify it !!!

val clear : ('a, 'b) t -> unit

Return the correspondance hashtable resp. from x to y and from y to x. Never modify it !!!

Empty a hash table. Use reset instead of clear to shrink the size of the bucket table to its initial size.

val reset : ('a, 'b) t -> unit

Empty a hash table. Use reset instead of clear to shrink the size of the bucket table to its initial size.

Empty a hash table and shrink the size of the bucket table to its initial size.

val create : int -> ('a, 'b) t

Empty a hash table and shrink the size of the bucket table to its initial size.

Create a new table, with the specified initial size

val add : ('a, 'b) t -> 'a -> 'b -> unit

Create a new table, with the specified initial size

Add a new binding to the table.

val y_of_x : ('a, 'b) t -> 'a -> 'b

Add a new binding to the table.

Association.

val x_of_y : ('a, 'b) t -> 'b -> 'a

Association.

Inverse association.

val removex : ('a, 'b) t -> 'a -> unit

Inverse association.

Remove a binding defined by its first element.

val removey : ('a, 'b) t -> 'b -> unit

Remove a binding defined by its first element.

Remove a binding defined by its second element.

val memx : ('a, 'b) t -> 'a -> bool

Remove a binding defined by its second element.

Is the object registered ?

val memy : ('a, 'b) t -> 'b -> bool

Is the object registered ?

Is the object registered ?

val iter : ('a, 'b) t -> ('a -> 'b -> unit) -> unit

Is the object registered ?

Iterate on bindings.

val fold : ('a, 'b) t -> 'c -> ('a -> 'b -> 'c -> 'c) -> 'c

Iterate on bindings.

Iterate on bindings and accumulating a result.

val cardinal : ('a, 'b) t -> int

Iterate on bindings and accumulating a result.

Return the number of bindings.

val print : ?first:(unit, Stdlib.Format.formatter, unit) Stdlib.format -> ?sep:(unit, Stdlib.Format.formatter, unit) Stdlib.format -> ?last:(unit, Stdlib.Format.formatter, unit) Stdlib.format -> ?firstbind:(unit, Stdlib.Format.formatter, unit) Stdlib.format -> ?sepbind:(unit, Stdlib.Format.formatter, unit) Stdlib.format -> ?lastbind:(unit, Stdlib.Format.formatter, unit) Stdlib.format -> (Stdlib.Format.formatter -> 'a -> unit) -> (Stdlib.Format.formatter -> 'b -> unit) -> Stdlib.Format.formatter -> ('a, 'b) t -> unit

Return the number of bindings.

Print the set of bindings.

module type Param = sig ... end

Input signature of the functor DHashhe.Make.

module type S = sig ... end

Output signature of the functor DHashhe.Make.

module Make (P : Param) : S with module HashX = P.HashX and module HashY = P.HashY

Functor building an implementation of the DHashtbl structure given two hashtables