Module Camllib.DMappe

Two-way map between two ordered data types

Functional semantics of operations. The generic interface assumess that types 'a and 'b should be comparable with the standard Pervasives.compare function.

type ('a, 'b) t

The type of two-way maps

val mapx : ('a, 'b) t -> ('a, 'b) Mappe.t

The type of two-way maps

val mapy : ('a, 'b) t -> ('b, 'a) Mappe.t

Return the correspondance map resp. from x to y and from y to x.

val empty : ('a, 'b) t

Return the correspondance map resp. from x to y and from y to x.

Empty map

val add : 'a -> 'b -> ('a, 'b) t -> ('a, 'b) t

Empty map

Add a new binding to the current map and return the new map.

val y_of_x : 'a -> ('a, 'b) t -> 'b

Add a new binding to the current map and return the new map.

Association

val x_of_y : 'b -> ('a, 'b) t -> 'a

Association

Inverse association

val remove : 'a -> ('a, 'b) t -> ('a, 'b) t

Inverse association

Remove a binding defined by its first element and return the new map.

val memx : 'a -> ('a, 'b) t -> bool

Remove a binding defined by its first element and return the new map.

Is the object in the map ?

val memy : 'b -> ('a, 'b) t -> bool

Is the object in the map ?

Is the object in the map ?

val merge : ('a, 'b) t -> ('a, 'b) t -> ('a, 'b) t

Is the object in the map ?

Merge the two double associations. If a key is bound to different data in the two arguments, raise Failure.

val common : ('a, 'b) t -> ('a, 'b) t -> ('a, 'b) t

Merge the two double associations. If a key is bound to different data in the two arguments, raise Failure.

Return the common bindings. If a key is bound to different data in the two arguments, raise Failure.

val intersetx : ('a, 'b) t -> 'a Sette.t -> ('a, 'b) t

Return the common bindings. If a key is bound to different data in the two arguments, raise Failure.

Select the two-way bindings x <->y with x in the set

val intersety : ('a, 'b) t -> 'b Sette.t -> ('a, 'b) t

Select the two-way bindings x <->y with x in the set

Select the two-way bindings x <->y with y in the set

val diffsetx : ('a, 'b) t -> 'a Sette.t -> ('a, 'b) t

Select the two-way bindings x <->y with y in the set

Remove the two-way bindings x <->y with x in the set

val diffsety : ('a, 'b) t -> 'b Sette.t -> ('a, 'b) t

Remove the two-way bindings x <->y with x in the set

Remove the two-way bindings x <->y with y in the set

val iter : ('a -> 'b -> unit) -> ('a, 'b) t -> unit

Remove the two-way bindings x <->y with y in the set

Iterate on bindings.

val fold : ('a -> 'b -> 'c -> 'c) -> ('a, 'b) t -> 'c -> 'c

Iterate on bindings.

Iterate on bindings and accumulating a result.

val setx : ('a, 'b) t -> 'a Sette.t

Iterate on bindings and accumulating a result.

Return the set of all objects in the first place of bindings.

val sety : ('a, 'b) t -> 'b Sette.t

Return the set of all objects in the first place of bindings.

Return the set of all objects in the second place of bindings.

val equalx : ('a, 'b) t -> ('a, 'b) t -> bool

Return the set of all objects in the second place of bindings.

val equaly : ('a, 'b) t -> ('a, 'b) t -> bool

val subsetx : ('a, 'b) t -> ('a, 'b) t -> bool

val subsety : ('a, 'b) t -> ('a, 'b) t -> bool

Test two two-way association for equality and inclusion. the x and y variants resp. exploit the tables x->y and y->x. For (slight) efficieny reason, one can choose one or the other variant.

val cardinal : ('a, 'b) t -> int

Test two two-way association for equality and inclusion. the x and y variants resp. exploit the tables x->y and y->x. For (slight) efficieny reason, one can choose one or the other variant.

Return the number of bindings.

val print : ?first:(unit, Stdlib.Format.formatter, unit) Stdlib.format -> ?sep:(unit, Stdlib.Format.formatter, unit) Stdlib.format -> ?last:(unit, Stdlib.Format.formatter, unit) Stdlib.format -> ?firstbind:(unit, Stdlib.Format.formatter, unit) Stdlib.format -> ?sepbind:(unit, Stdlib.Format.formatter, unit) Stdlib.format -> ?lastbind:(unit, Stdlib.Format.formatter, unit) Stdlib.format -> (Stdlib.Format.formatter -> 'a -> unit) -> (Stdlib.Format.formatter -> 'b -> unit) -> Stdlib.Format.formatter -> ('a, 'b) t -> unit

Return the number of bindings.

Print the set of bindings.

module type Param = sig ... end

Input signature of the functor DMappe.Make.

module type S = sig ... end

Output signature of the functor DMappe.Make.

module Make (P : Param) : S with module MappeX = P.MappeX and module MappeY = P.MappeY

Functor building an implementation of the DMappe structure given two map structures.