Parameter Make.T

module MapV : Mappe.S

Map module for associating attributes to vertices, of type MapV.key

module MapE : Mappe.S with type key = MapV.key * MapV.key

Map module for associating attributes to vertices, of type MapV.key