Module Camllib.Parse

Linking lexer and parser function, with (basic) error messages

exception Lex_error

val lex_eol : Stdlib.Lexing.lexbuf -> unit

Function to call on line returns in lexer

val parse_lexbuf : lexer:(Stdlib.Lexing.lexbuf -> 'token) -> parser:((Stdlib.Lexing.lexbuf -> 'token) -> Stdlib.Lexing.lexbuf -> 'ast) -> Stdlib.Lexing.lexbuf -> 'ast

Takes as input a lexer, a parser compatible with the lexer (ie, they share the same type for tokens), and a lexbuf,

Returns the AST built by the parser.

val parse_string : lexer:(Stdlib.Lexing.lexbuf -> 'token) -> parser:((Stdlib.Lexing.lexbuf -> 'token) -> Stdlib.Lexing.lexbuf -> 'ast) -> string -> 'ast

Same as before, but create itself a lexbuf from the given string.

val parse_file : lexer:(Stdlib.Lexing.lexbuf -> 'token) -> parser:((Stdlib.Lexing.lexbuf -> 'token) -> Stdlib.Lexing.lexbuf -> 'ast) -> string -> 'ast

Same as before, but create itself a lexbuf from the given filename.