Module Camllib.Print

Printing functions using module Format

Printing functions for standard datatypes (lists,arrays,...)

In the following functions, optional arguments ?first, ?sep, ?last denotes the formatting instructions (under the form of a format string) issued at the beginning, between two elements, and at the end.

The functional argument(s) indicate(s) how to print elements.

val list : ?first:(unit, Stdlib.Format.formatter, unit) Stdlib.format -> ?sep:(unit, Stdlib.Format.formatter, unit) Stdlib.format -> ?last:(unit, Stdlib.Format.formatter, unit) Stdlib.format -> (Stdlib.Format.formatter -> 'a -> unit) -> Stdlib.Format.formatter -> 'a list -> unit

Print a list

val array : ?first:(unit, Stdlib.Format.formatter, unit) Stdlib.format -> ?sep:(unit, Stdlib.Format.formatter, unit) Stdlib.format -> ?last:(unit, Stdlib.Format.formatter, unit) Stdlib.format -> (Stdlib.Format.formatter -> 'a -> unit) -> Stdlib.Format.formatter -> 'a array -> unit

Print an array

val pair : ?first:(unit, Stdlib.Format.formatter, unit) Stdlib.format -> ?sep:(unit, Stdlib.Format.formatter, unit) Stdlib.format -> ?last:(unit, Stdlib.Format.formatter, unit) Stdlib.format -> (Stdlib.Format.formatter -> 'a -> unit) -> (Stdlib.Format.formatter -> 'b -> unit) -> Stdlib.Format.formatter -> ('a * 'b) -> unit

Print a pair

val option : ?first:(unit, Stdlib.Format.formatter, unit) Stdlib.format -> ?last:(unit, Stdlib.Format.formatter, unit) Stdlib.format -> (Stdlib.Format.formatter -> 'a -> unit) -> Stdlib.Format.formatter -> 'a option -> unit

Print an optional element

val hash : ?first:(unit, Stdlib.Format.formatter, unit) Stdlib.format -> ?sep:(unit, Stdlib.Format.formatter, unit) Stdlib.format -> ?last:(unit, Stdlib.Format.formatter, unit) Stdlib.format -> ?firstbind:(unit, Stdlib.Format.formatter, unit) Stdlib.format -> ?sepbind:(unit, Stdlib.Format.formatter, unit) Stdlib.format -> ?lastbind:(unit, Stdlib.Format.formatter, unit) Stdlib.format -> (Stdlib.Format.formatter -> 'a -> unit) -> (Stdlib.Format.formatter -> 'b -> unit) -> Stdlib.Format.formatter -> ('a, 'b) Stdlib.Hashtbl.t -> unit

Print an hashtable

val weak : ?first:(unit, Stdlib.Format.formatter, unit) Stdlib.format -> ?sep:(unit, Stdlib.Format.formatter, unit) Stdlib.format -> ?last:(unit, Stdlib.Format.formatter, unit) Stdlib.format -> (Stdlib.Format.formatter -> 'a -> unit) -> Stdlib.Format.formatter -> 'a Stdlib.Weak.t -> unit

Print a weak pointer array

Useful functions

val string_of_print : (Stdlib.Format.formatter -> 'a -> unit) -> 'a -> string

Transforms a printing function into a conversion-to-string function.

val print_of_string : ('a -> string) -> Stdlib.Format.formatter -> 'a -> unit

Transforms a conversion-to-string function to a printing function.

val sprintf : ?margin:int -> ('a, Stdlib.Format.formatter, unit, string) Stdlib.format4 -> 'a

Better sprintf function than Format.sprintf, as it takes the same kind of formatters as other Format.Xprintf functions.

val escaped : ?linebreak:char -> string -> string

Escape a string, replacing line breaks by linebreak (default '\n'). When used for DOT output, '\l' and '\r' produces respectively left or right justified lines, instead of center justified lines.