Module SetList.Make

Functor building an implementation of the SetList structure given a totally ordered type.

Parameters

module Ord : Stdlib.Set.OrderedType

Signature

type elt = Ord.t

The type of the set elements.

type t

The type of the set elements.

The type of sets.

val print : ?first:(unit, Stdlib.Format.formatter, unit) Stdlib.format -> ?sep:(unit, Stdlib.Format.formatter, unit) Stdlib.format -> ?last:(unit, Stdlib.Format.formatter, unit) Stdlib.format -> (Stdlib.Format.formatter -> elt -> unit) -> Stdlib.Format.formatter -> t -> unit

The type of sets.

Printing function

val empty : t

Printing function

The empty set.

val is_empty : t -> bool

The empty set.

Test whether a set is empty or not.

val mem : elt -> t -> bool

Test whether a set is empty or not.

mem x s tests whether x belongs to the set s.

val of_list : elt list -> t

mem x s tests whether x belongs to the set s.

Conversion from a list (unsafe operation)

val to_list : t -> elt list

Conversion from a list (unsafe operation)

Conversion to a list

val singleton : elt -> t

Conversion to a list

singleton x returns the one-element set containing only x.

val add : elt -> t -> t

singleton x returns the one-element set containing only x.

add x s returns a set containing all elements of s, plus x. If x was already in s, s is returned unchanged.

val remove : elt -> t -> t

add x s returns a set containing all elements of s, plus x. If x was already in s, s is returned unchanged.

remove x s returns a set containing all elements of s, except x. If x was not in s, s is returned unchanged.

val union : t -> t -> t

remove x s returns a set containing all elements of s, except x. If x was not in s, s is returned unchanged.

val inter : t -> t -> t

val diff : t -> t -> t

Union, intersection and set difference.

val compare : t -> t -> int

Total ordering between sets. Can be used as the ordering function for doing sets of sets.

val equal : t -> t -> bool

Total ordering between sets. Can be used as the ordering function for doing sets of sets.

equal s1 s2 tests whether the sets s1 and s2 are equal, that is, contain equal elements.

val subset : t -> t -> bool

equal s1 s2 tests whether the sets s1 and s2 are equal, that is, contain equal elements.

subset s1 s2 tests whether the set s1 is a subset of the set s2.

val iter : (elt -> unit) -> t -> unit

subset s1 s2 tests whether the set s1 is a subset of the set s2.

iter f s applies f in turn to all elements of s. The order in which the elements of s are presented to f is in ascending order.

val fold : (elt -> 'a -> 'a) -> t -> 'a -> 'a

iter f s applies f in turn to all elements of s. The order in which the elements of s are presented to f is in ascending order.

fold f s a computes (f xN ... (f x2 (f x1 a))...), where x1 ... xN are the elements of s. The order in which elements of s are presented to f is in ascending order.

parameter f
function

parameter s
set

parameter a
accumulator

returns
the computed accumulator

raises Not_found
if no fount

val fold_right : (elt -> 'a -> 'a) -> t -> 'a -> 'a

fold f s a computes (f xN ... (f x2 (f x1 a))...), where x1 ... xN are the elements of s. The order in which elements of s are presented to f is in ascending order.

parameter f
function

parameter s
set

parameter a
accumulator

returns
the computed accumulator

raises Not_found
if no fount

val fold_left : ('a -> elt -> 'a) -> 'a -> t -> 'a

Idem as List.fold_X functions

val cardinal : t -> int

Idem as List.fold_X functions

Return the number of elements of a set.

val elements : t -> elt list

Return the number of elements of a set.

Return the list of all elements of the given set. The returned list is sorted in increasing order.

val min_elt : t -> elt

Return the list of all elements of the given set. The returned list is sorted in increasing order.

Return the smallest element of the given set (with respect to the Ord.compare ordering), or raise Not_found if the set is empty.

val max_elt : t -> elt

Return the smallest element of the given set (with respect to the Ord.compare ordering), or raise Not_found if the set is empty.

Same as min_elt, but returns the largest element of the given set.

val choose : t -> elt

Same as min_elt, but returns the largest element of the given set.

Return one element of the given set, or raise Not_found if the set is empty. Which element is chosen is unspecified, but equal elements will be chosen for equal sets.

val filter : (elt -> bool) -> t -> t

Return one element of the given set, or raise Not_found if the set is empty. Which element is chosen is unspecified, but equal elements will be chosen for equal sets.

filter p s returns the set of all elements in s that satisfy predicate p.

val partition : (elt -> bool) -> t -> t * t

filter p s returns the set of all elements in s that satisfy predicate p.

partition p s returns a pair of sets (s1, s2), where s1 is the set of all the elements of s that satisfy the predicate p, and s2 is the set of all the elements of s that do not satisfy p.

val exists : (elt -> bool) -> t -> bool

partition p s returns a pair of sets (s1, s2), where s1 is the set of all the elements of s that satisfy the predicate p, and s2 is the set of all the elements of s that do not satisfy p.

exists p s checks if at least one element of the set satisfies the predicate p.

val for_all : (elt -> bool) -> t -> bool

exists p s checks if at least one element of the set satisfies the predicate p.

for_all p s checks if all elements of the set satisfy the predicate p.