Module Sette.Compare

val split : ('a -> 'a -> int) -> 'a -> 'a t -> 'a t * bool * 'a t

Meant to be internal, but exporting needed for Mappe.maptoset.

val add : ('a -> 'a -> int) -> 'a -> 'a t -> 'a t

Meant to be internal, but exporting needed for Mappe.maptoset.

val mem : ('a -> 'a -> int) -> 'a -> 'a t -> bool

val remove : ('a -> 'a -> int) -> 'a -> 'a t -> 'a t

val union : ('a -> 'a -> int) -> 'a t -> 'a t -> 'a t

val inter : ('a -> 'a -> int) -> 'a t -> 'a t -> 'a t

val diff : ('a -> 'a -> int) -> 'a t -> 'a t -> 'a t

val equal : ('a -> 'a -> int) -> 'a t -> 'a t -> bool

val compare : ('a -> 'a -> int) -> 'a t -> 'a t -> int

val subset : ('a -> 'a -> int) -> 'a t -> 'a t -> bool

val filter : ('a -> 'a -> int) -> ('a -> bool) -> 'a t -> 'a t

val partition : ('a -> 'a -> int) -> ('a -> bool) -> 'a t -> 'a t * 'a t