Module Camllib.Symbol

Symbol table, for string management

type t

Type of symbols (actually integers)

val add : string -> t

Returns the symbol associated to the given string, after having possibly registered the string if it wasn't (in this case, the symbol is fresh symbol).

val exists : string -> bool

Returns the symbol associated to the given string, after having possibly registered the string if it wasn't (in this case, the symbol is fresh symbol).

Is the string already registered ?

val of_string : string -> t

Is the string already registered ?

Returns the existing symbol associated to the (registered) string. Raises Not_found otherwise.

val to_string : t -> string

Returns the existing symbol associated to the (registered) string. Raises Not_found otherwise.

Returns the string represented by the symbol.

val print : Stdlib.Format.formatter -> t -> unit

Returns the string represented by the symbol.

Prints the symbol (its associated string).

val equal : t -> t -> bool

Prints the symbol (its associated string).

Equality test

val compare : t -> t -> int

Equality test

Comparison (do not correspond at all to alphabetic order, depend on the registration order of names in the module)

module HashedType : Stdlib.Hashtbl.HashedType with type t = t

To use hashtables on type t

module OrderedType : Stdlib.Set.OrderedType with type t = t

To use hashtables on type t

module Hash : Stdlib.Hashtbl.S with type key = t

Hashtables on type t

module Set : Sette.S with type elt = t

Hashtables on type t

module Map : Mappe.S with type key = t and module Setkey = Set

Sets on type t

val print_set : Stdlib.Format.formatter -> Set.t -> unit

Prints sets of symbols.

val print_hash : (Stdlib.Format.formatter -> 'a -> unit) -> Stdlib.Format.formatter -> 'a Hash.t -> unit

Prints sets of symbols.

Prints hashtables on symbols.

val print_map : (Stdlib.Format.formatter -> 'a -> unit) -> Stdlib.Format.formatter -> 'a Map.t -> unit

Prints hashtables on symbols.

Prints maps on symbols.