Module Camllib.Union

Union-find Abstract Data Types

type 'a t

The type of the data structure storing set membership (the universe)

val create : int -> 'a t

Create a new universe set

val add : 'a t -> 'a -> unit

Create a new universe set

Add an element to the universe (initially belonging to its singleton)

val find : 'a t -> 'a -> 'a

Add an element to the universe (initially belonging to its singleton)

val union : 'a t -> 'a -> 'a -> 'a

Computes the union of two sets and returns the resulting set

val extract : 'a t -> 'a list list

Computes the union of two sets and returns the resulting set

Extract the list of sets